--- test - oldtests/api-v1/MaruResults.json
@@ -1,44 +1,0 @@
-[{
-    "sourceType": "solidity-file",
-    "sourceFormat": "text",
-    "sourceList": [
-        "/home/rocky/truffle/mythjs-should-detect/bernhard_medium_post/contracts/Migrations.sol"
-    ],
-    "issues": [
-        {
-            "locations": [
-                {
-                    "sourceMap": "0:23:0"
-                }
-            ],
-            "swcID": "SWC-103",
-            "swcTitle": "Floating Pragma",
-            "description": {
-                "head": "A floating pragma is set.",
-                "tail": "It is recommended to make a conscious choice on what version of Solidity is used for compilation. Currently any version equal or grater than \"0.5.0\" is allowed."
-            },
-            "severity": "Low",
-            "extra": {}
-        },
-        {
-            "locations": [
-                {
-                    "sourceMap": "400:19:0"
-                }
-            ],
-            "swcID": "SWC-109",
-            "swcTitle": "Uninitialized Storage Pointer",
-            "description": {
-                "head": "Dangerous use of uninitialized storage variables.",
-                "tail": "Uninitialized storage variables of user-defined type can point to unexpected storage locations. Initialize variable \"upgraded\" or set the storage attribute \"memory\"."
-            },
-            "severity": "Low",
-            "extra": {}
-        }
-    ],
-    "meta": {
-        "selected_compiler": "0.5.0",
-        "error": [],
-        "warning": []
-    }
-}]
