--- test - oldtests/insert-into-truffle-configs/test1/truffle-config.js
@@ -1,2 +1,0 @@
-module.exports = {
-}
