--- test - oldtests/insert-into-truffle-configs/test2/truffle-config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  plugins: [],
-}
