--- test - oldtests/insert-into-truffle-configs/test4/truffle-config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  plugins: [
-  ]
-}
