--- test - oldtests/insert-into-truffle-configs/test5/truffle-config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  plugins: [
-  ],
-}
