--- test - oldtests/sample-truffle/metacoin/contracts/ConvertLib.sol
@@ -1,8 +1,0 @@
-pragma solidity >=0.4.22 <0.6.0;
-
-library ConvertLib{
-	function convert(uint amount,uint conversionRate) public pure returns (uint convertedAmount)
-	{
-		return amount * conversionRate;
-	}
-}
