--- test - oldtests/sample-truffle/metacoin/truffle-config.js
@@ -1,20 +1,0 @@
-module.exports = {
-  networks: {
-    development: {
-      host: "localhost",
-      port: 8545,
-      network_id: "*", // Match any network id
-      gas: 5000000
-    }
-  },
-  compilers: {
-    solc: {
-      settings: {
-        optimizer: {
-          enabled: true, // Default: false
-          runs: 200      // Default: 200
-        },
-      }
-    }
-  }
-};
