--- test - oldtests/sample-truffle/sample_import/.gitignore
@@ -1,2 +1,0 @@
-/build
-/migrations
