--- test - oldtests/sample-truffle/simple_dao/truffle.js
@@ -1,26 +1,0 @@
-/*
- * NB: since truffle-hdwallet-provider 0.0.5 you must wrap HDWallet providers in a
- * function when declaring them. Failure to do so will cause commands to hang. ex:
- * ```
- * mainnet: {
- *     provider: function() {
- *       return new HDWalletProvider(mnemonic, 'https://mainnet.infura.io/<infura-key>')
- *     },
- *     network_id: '1',
- *     gas: 4500000,
- *     gasPrice: 10000000000,
- *   },
- */
-
-module.exports = {
-    // See <http://truffleframework.com/docs/advanced/configuration>
-    // to customize your Truffle configuration!
-    compilers: {
-        solc: {
-            version: "0.4.25"  // ex:  "0.4.20". (Default: Truffle's installed solc)
-        }
-    },
-    plugins: [
-        'truffle-security'
-    ]
-};
