--- test - oldtests/sample-truffle/storage_arrays/contracts/PublicStorageArray.sol
@@ -1,5 +1,0 @@
-pragma solidity ^0.4.22;
-
-contract PublicStorageArray {
-    bytes32[] public states = [bytes32(0)];
-}
