+++ test/test_apiclients.js
@@ -7,7 +7,7 @@ const armlet = require('armlet');
 const sinon = require('sinon');
 const trufstuf = require('../lib/trufstuf');
 const mythx = require('../lib/mythx');
-const rewiredHelpers = rewire('../helpersRefactor');
+const rewiredHelpers = rewire('../helpers');
 const util = require('util');
 const yaml = require('js-yaml');
 const armletClass = require('../classes/armlet');
@@ -107,6 +107,96 @@ describe('API Client Classes', function() {
             assert.equal(results.errors.length, 0);
             assert.equal(results.objects.length, 1);
         });
+
+        it('should return 0 mythXIssues objects and 1 error', async function() {
+          const config = {
+              _: [],
+              debug: true,
+              logger: {debug: debuggerStub},
+              style: 'test-style',
+              progress: false,
+          }
+          const jsonFile = `${__dirname}/sample-truffle/simple_dao/build/mythx/contracts/simple_dao.json`;
+          const simpleDaoJSON = await util.promisify(fs.readFile)(jsonFile, 'utf8');
+          const contracts = mythx.newTruffleObjToOldTruffleByContracts(JSON.parse(simpleDaoJSON));
+          const objContracts = [ { contractName: "SimpleDAO", contract: contracts[0] } ];
+          const mythXInput = mythx.truffle2MythXJSON(objContracts[0].contract);
+          doAnalysisFromClientStub.resolves({
+            issues: [],
+            status: { status: 'Error'},
+          });
+
+          //pathStub.resolve.returns("/build/contracts/mythx/contracts/contract.sol");
+          apiClient = new APIClient(config, "truffle");
+          const results = await apiClient.doAnalysis(objContracts);
+          mythXInput.analysisMode = 'quick';
+          assert.ok(doAnalysisFromClientStub.calledWith({
+              clientToolName: 'truffle',
+              data: mythXInput,
+              noCacheLookup: false,
+          }, 300000, undefined));
+          assert.equal(results.errors.length, 1);
+          assert.equal(results.objects.length, 0);
+      });
+
+
+      it('should return 1 mythXIssues object and 1 error', async function() {
+        const config = {
+            _: [],
+            debug: true,
+            logger: {debug: debuggerStub},
+            style: 'test-style',
+            progress: false,
+        }
+        const jsonFile = `${__dirname}/sample-truffle/simple_dao/build/mythx/contracts/simple_dao.json`;
+        const simpleDaoJSON = await util.promisify(fs.readFile)(jsonFile, 'utf8');
+        const contracts = mythx.newTruffleObjToOldTruffleByContracts(JSON.parse(simpleDaoJSON));
+        const objContracts = [ { contractName: "SimpleDAO", contract: contracts[0] } ];
+        const mythXInput = mythx.truffle2MythXJSON(objContracts[0].contract);
+        doAnalysisFromClientStub.onFirstCall().resolves({
+            issues: {},
+            status: { status: 'Error' },
+        });
+        doAnalysisFromClientStub.onSecondCall().resolves({
+          issues: [{
+              'sourceFormat': 'evm-byzantium-bytecode',
+              'sourceList': [
+                  `${__dirname}/sample-truffle/simple_dao/build/mythx/contracts/simple_dao.json`
+              ],
+              'sourceType': 'raw-bytecode',
+              'issues': [{
+                  'description': {
+                      'head': 'Head message',
+                      'tail': 'Tail message'
+                  },
+                  'locations': [{
+                      'sourceMap': '444:1:0'
+                  }],
+                  'severity': 'High',
+                  'swcID': 'SWC-000',
+                  'swcTitle': 'Test Title'
+              }],
+              'meta': {
+                  'selected_compiler': '0.5.0',
+                  'error': [],
+                  'warning': []
+              },
+          }],
+          status: {status: 'Pending' },
+      });
+        //pathStub.resolve.returns("/build/contracts/mythx/contracts/contract.sol");
+        apiClient = new APIClient(config, "truffle");
+        const results = await apiClient.doAnalysis(objContracts);
+        mythXInput.analysisMode = 'quick';
+        assert.ok(doAnalysisFromClientStub.calledWith({
+            clientToolName: 'truffle',
+            data: mythXInput,
+            noCacheLookup: false,
+        }, 300000, undefined));
+        assert.equal(results.errors.length, 1);
+        assert.equal(results.objects.length, 0);
+    });
+
 
     })
 });
